+++ src/types/bar.ts
@@ -1,9 +1,17 @@
-import { EChartOption } from 'echarts';
+import { BarSeriesOption } from 'echarts';
+
+import { SeriesType } from './visual-editor';
 
 /**
  * Bar Series Options
  */
-export interface BarSeriesOptions extends EChartOption.SeriesLine {
+export interface BarSeriesOptions extends BarSeriesOption {
+  /**
+   * Type
+   * type?: string in BarSeriesOption not working as expected with SeriesByType
+   */
+  type: SeriesType.BAR;
+
   /**
    * Encode
    */