+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Change Log
 
-## 7.0.0 (IN PROGRESS)
+## 6.6.0 (2025-02-20)
 
 ### Features / Enhancements
 
 - Added E2E tests (#349)
-- Update packages for Code Editor (#352)
+- Updated packages for Code Editor (#352)
+- Updated Apache ECharts 5.6.0 and dependencies (#355)
 
 ## 6.5.0 (2024-10-25)